Refactored code to not use a fragment for authorization.
